Move the network tethering parts into conditional blocks to make the whole (non-working thing) optional. Start de-hardcoding stuff like the disc image file path and auto-detect enabling cdrom mode or not depending on the .ISO extension.

+++ _boot_pc_android_mass_storage_enable_root.sh
@@ -15,6 +15,10 @@
 
 #  `mount | grep configfs` to get the mount point/path for the virtual config filesystem
 
+
+SWY_MOUNT_FILE='/sdcard/Download/netboot.xyz.iso'
+SWY_TETHER=0
+
 # swy: usually mounted at /config
 CONFIGFS=`mount -t configfs | head -n1 | cut -d' ' -f 3`
 
@@ -39,11 +43,11 @@ mkdir configs/swyconfig.1/strings/0x409
 echo "first rndis, then mass_storage to work on win32" > configs/swyconfig.1/strings/0x409/configuration
 
 # --
-
-# swy: add a RNDIS Windows USB tethering function, here we seem to need the suffix
-mkdir functions/gsi.rndis
-ln -s functions/gsi.rndis configs/swyconfig.1 # swy: add a symbolic link to put our function into a premade config folder
-
+if [ $SWY_TETHER ]; then
+  # swy: add a RNDIS Windows USB tethering function, here we seem to need the suffix
+  mkdir functions/gsi.rndis
+  ln -s functions/gsi.rndis configs/swyconfig.1 # swy: add a symbolic link to put our function into a premade config folder
+fi
 # --
 
 mkdir functions/mass_storage.0 # swy: create a gadget function of type 'mass_storage', only the part after the . is customizable
@@ -53,8 +57,11 @@ mkdir functions/mass_storage.0 # swy: create a gadget function of type 'mass_sto
 #      if the drive appears blank/0 bytes there's your problem.
 echo "y"                                   > functions/mass_storage.0/lun.0/ro
 echo "y"                                   > functions/mass_storage.0/lun.0/removable
-echo "y"                                   > functions/mass_storage.0/lun.0/cdrom
-echo "/sdcard/Download/netboot.xyz.iso"    > functions/mass_storage.0/lun.0/file # swy: make sure we assign the actual path last, or setting ro/cdrom won't work until we empty this
+case "$SWY_MOUNT_FILE" in
+*.iso) echo echo "y"                       > functions/mass_storage.0/lun.0/cdrom
+*    ) echo echo "n"                       > functions/mass_storage.0/lun.0/cdrom
+esac
+echo "$SWY_MOUNT_FILE"                     > functions/mass_storage.0/lun.0/file # swy: make sure we assign the actual path last, or setting ro/cdrom won't work until we empty this
 
 ln -s functions/mass_storage.0 configs/swyconfig.1 # swy: add a symbolic link to put our function into a premade config folder
 
@@ -67,45 +74,52 @@ echo "" > ../g1/UDC
 getprop sys.usb.controller > UDC
 setprop sys.usb.state mass_storage
 
+# --
+if [ $SWY_TETHER ]; then
+  ip address add 10.20.30.1/24 dev rndis0
+  ip link set rndis0 up
 
-ip address add 10.20.30.1/24 dev rndis0
-ip link set rndis0 up
+  # swy: https://github.com/luftreich/android-wired-tether/blob/725e79e9/native/tether/tetherStartStop.cpp#L154
+  # swy: these don't seem to work well, networking is broken at both ends
+  iptables -F
+  iptables -F -t nat
+  iptables -I FORWARD -m state --state ESTABLISHED,RELATED -j ACCEPT
+  iptables -I FORWARD -s 10.20.30.0/24 -j ACCEPT
+  iptables -P FORWARD DROP
+  iptables -t nat -I POSTROUTING -s 10.20.30.0/24 -j MASQUERADE
 
-# swy: https://github.com/luftreich/android-wired-tether/blob/725e79e9/native/tether/tetherStartStop.cpp#L154
-# swy: these don't seem to work well, networking is broken at both ends
-iptables -F
-iptables -F -t nat
-iptables -I FORWARD -m state --state ESTABLISHED,RELATED -j ACCEPT
-iptables -I FORWARD -s 10.20.30.0/24 -j ACCEPT
-iptables -P FORWARD DROP
-iptables -t nat -I POSTROUTING -s 10.20.30.0/24 -j MASQUERADE
+  # swy: these get added by Android, found while diff'ing
+  iptables -A tetherctrl_FORWARD -j bw_global_alert
+  iptables -A tetherctrl_FORWARD -i wlan0 -o rndis0 -m state --state RELATED,ESTABLISHED -g tetherctrl_counters
+  iptables -A tetherctrl_FORWARD -i rndis0 -o wlan0 -m state --state INVALID -j DROP
+  iptables -A tetherctrl_FORWARD -i rndis0 -o wlan0 -g tetherctrl_counters
 
-# swy: these get added by Android, found while diff'ing
-iptables -A tetherctrl_FORWARD -j bw_global_alert
-iptables -A tetherctrl_FORWARD -i wlan0 -o rndis0 -m state --state RELATED,ESTABLISHED -g tetherctrl_counters
-iptables -A tetherctrl_FORWARD -i rndis0 -o wlan0 -m state --state INVALID -j DROP
-iptables -A tetherctrl_FORWARD -i rndis0 -o wlan0 -g tetherctrl_counters
-
-# swy: --port=0 disables the DNS functionality, we only want it to work as a DHCP server
-#      --enable-tftp --tftp-root="/sdcard/Download" (seemingly no tftp support in the bundled version ¿?)
-#      --conf-file=/data/tmp/dnsmasq.conf
-killall -9 dnsmasq
-dnsmasq --no-daemon --no-hosts --no-resolv --server=8.8.8.8 --interface=rndis0 --dhcp-range=tether,10.20.30.2,10.20.30.20,1h
+  # swy: --port=0 disables the DNS functionality, we only want it to work as a DHCP server
+  #      --enable-tftp --tftp-root="/sdcard/Download" (seemingly no tftp support in the bundled version ¿?)
+  #      --conf-file=/data/tmp/dnsmasq.conf
+  killall -9 dnsmasq
+  dnsmasq --no-daemon --no-hosts --no-resolv --server=8.8.8.8 --interface=rndis0 --dhcp-range=tether,10.20.30.2,10.20.30.20,1h
+fi
+# --
 
 echo "[i] press any key to exit the mass storage gadget mode..." && read
 echo 
 
-# swy: tear down the tables
-iptables -F
-iptables -t nat -F
-iptables -X
-iptables -t nat -X
-iptables -P FORWARD ACCEPT
+# --
+if [ $SWY_TETHER ]; then
+  # swy: tear down the tables
+  iptables -F
+  iptables -t nat -F
+  iptables -X
+  iptables -t nat -X
+  iptables -P FORWARD ACCEPT
 
 
-killall -9 dnsmasq
-ip link set rndis0 down
-ip address delete 10.20.30.1/32 dev rndis0
+  killall -9 dnsmasq
+  ip link set rndis0 down
+  ip address delete 10.20.30.1/32 dev rndis0
+fi
+# --
 
 # swy: detach the gadget from the physical USB port
 echo "" > UDC
@@ -116,7 +130,9 @@ rmdir configs/swyconfig.1/strings/0x409  #swy: deallocate the configuration stri
 rmdir configs/swyconfig.1/               #swy: now we can remove the empty config
 
 rmdir functions/mass_storage.0           #swy: remove the now-unlinked function
-rmdir functions/gsi.rndis                #swy: remove the now-unlinked function
+if [ $SWY_TETHER ]; then
+  rmdir functions/gsi.rndis              #swy: remove the now-unlinked function
+fi
 
 rmdir strings/0x409                      #swy: deallocate the gadget strings
 cd .. && rmdir swy                       #swy: remove the now-empty gadget
